Docs : resson폴더 삭제 ws웹소켓 서버 연결
