--- resson/default/private_key
@@ -1,1 +1,0 @@
-75cd1eb7e78a4e458185246fa20c2216583d4dc963130693ade6552afba34cb1
