--- resson/test.js
@@ -1,7 +1,0 @@
-function init() {
-  const single_q = "hello world";
-  const double_q = "hello world";
-  const no_semi = 500;
-  console.log("hello world!");
-}
-init();
